Move sqs stuff to grapl-common. Begin preparing engagement-creator for de-lambdification
