+++ src/python/analyzer_executor/src/analyzer_executor_lib/analyzer_executor.py
@@ -27,7 +27,7 @@ from typing import (
 
 import boto3
 from analyzer_executor_lib.redis_cache import EitherCache, construct_redis_client
-from analyzer_executor_lib.sqs_types import S3PutRecordDict, SQSMessageBody
+from grapl_common.sqs.sqs_types import S3PutRecordDict, SQSMessageBody
 from grapl_analyzerlib.analyzer import Analyzer
 from grapl_analyzerlib.execution import ExecutionComplete, ExecutionFailed, ExecutionHit
 from grapl_analyzerlib.grapl_client import GraphClient
