+++ src/python/analyzer_executor/src/analyzer_executor_lib/event_retriever.py
@@ -5,7 +5,7 @@ import traceback
 from typing import Iterator
 
 import boto3
-from analyzer_executor_lib.sqs_types import SQSMessage
+from grapl_common.sqs.sqs_types import SQSMessage
 from grapl_common.env_helpers import SQSClientFactory
 from grapl_common.grapl_logger import get_module_grapl_logger
