+++ src/python/engagement-creator/BUILD
@@ -1,8 +1,1 @@
 python_library()
-
-python_awslambda(
-    name="lambda",
-    output_path="engagement-creator-lambda.zip",
-    runtime="python3.7",
-    handler="engagement_creator.py:lambda_handler",
-)
