+++ src/python/grapl-common/grapl_common/sqs/sqs_timeout_manager.py
@@ -4,7 +4,7 @@ import asyncio
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Coroutine
 
-from analyzer_executor_lib.sqs_types import SQSMessageId, SQSReceiptHandle
+from grapl_common.sqs.sqs_types import SQSMessageId, SQSReceiptHandle
 from grapl_common.grapl_logger import get_module_grapl_logger
 
 if TYPE_CHECKING:
